+++ lib/typhoeus/response/header.rb
@@ -6,7 +6,7 @@ module Typhoeus
     # Values can be strings (normal case) or arrays of strings (for duplicates headers)
     #
     # @api private
-    class Header < Hash
+    class Header < DelegateClass(Hash)
 
       # Create a new header.
       #
@@ -15,10 +15,14 @@ module Typhoeus
       #
       # @param [ String ] raw The raw header.
       def initialize(raw)
+        super({})
         @raw = raw
         @sanitized = {}
         parse
-        set_default_proc_on(self, lambda { |h, k| @sanitized[k.to_s.downcase] })
+      end
+
+      def [](key)
+        fetch(key) { @sanitized[key.to_s.downcase] }
       end
 
       # Parses the raw header.
@@ -32,8 +36,9 @@ module Typhoeus
             process_pair(k, v)
           end
         when String
-          raw.lines.each do |header|
-            next if header.empty? || header.start_with?( 'HTTP/1.' )
+          raw.split(/\r?\n(?!\s)/).each do |header|
+            header.strip!
+            next if header.empty? || header.start_with?( 'HTTP/' )
             process_line(header)
           end
         end
@@ -46,7 +51,7 @@ module Typhoeus
       # @return [ void ]
       def process_line(header)
         key, value = header.split(':', 2)
-        process_pair(key.strip, value.strip)
+        process_pair(key.strip, (value ? value.strip.gsub(/\r?\n\s*/, ' ') : ''))
       end
 
       # Sets key value pair for self and @sanitized.